+++ ui/js/hud/WindPanel.js
@@ -8,12 +8,13 @@ import { DisplayUnit } from "../DisplayUnit.js";
 import { getWindBarb } from "../WindBarb.js";
 
 export const WindPanel = L.Control.extend({
-  options: { position: "topleft" },
+  options: { position: "bottomleft" },
 
   onAdd: function () {
     const container = L.DomUtil.create("div", "windBarbControl leaflet-bar");
     L.DomEvent.disableClickPropagation(container);
     container.id = "windBarbUI";
+    container.style.display = "none";
     container.innerHTML = `
       <div><b>Wind</b></div>
       <div id="windBarbContainer"></div>
@@ -83,8 +84,10 @@ export const WindPanel = L.Control.extend({
     //if we don't have the right data, hide ourself.
     if (!state.aws || !state.twa)
       this.hide();
-    else
+    else {
       this.setSpeed(state.aws, state.twa);
+      this.show();
+    }
   },
 
   clearSpeed: function () {